Refactor inline images PR for project conventions
