+++ frontend/apps/main/src/features/conversation/message/attachment/MessageAttachmentPreview.vue
@@ -6,14 +6,9 @@
       class="flex items-center cursor-pointer"
     >
       <div>
-        <ImageAttachmentPreview
-          v-if="isImage(attachment)"
-          :attachment="attachment"
-          @preview="openLightbox"
-        />
         <div
-          v-else-if="isAudio(attachment)"
-          class="flex items-center gap-2 rounded-lg border bg-muted/40 px-3 py-2"
+          v-if="isAudio(attachment)"
+          class="flex items-center gap-2 rounded border bg-muted/40 px-3 py-2"
         >
           <audio controls preload="auto" class="h-8 max-w-[260px]">
             <source :src="attachment.url" />
@@ -22,14 +17,14 @@
             :href="attachment.url"
             download
             class="p-1 rounded hover:bg-muted shrink-0"
-            :title="t('imageLightbox.download')"
-            :aria-label="t('imageLightbox.download')"
+            :title="t('globals.terms.download')"
+            :aria-label="t('globals.terms.download')"
             @click.stop
           >
             <Download class="w-4 h-4 text-muted-foreground" />
           </a>
         </div>
-        <FileAttachmentPreview v-else :attachment="attachment" />
+        <AttachmentItem v-else :attachment="attachment" @preview="openLightbox" />
       </div>
     </div>
   </div>
@@ -45,8 +40,7 @@
 import { ref, computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { Download } from 'lucide-vue-next'
-import ImageAttachmentPreview from '@/features/conversation/message/attachment/ImageAttachmentPreview.vue'
-import FileAttachmentPreview from '@/features/conversation/message/attachment/FileAttachmentPreview.vue'
+import AttachmentItem from '@/features/conversation/message/attachment/AttachmentItem.vue'
 import ImageLightbox from '@/components/ImageLightbox.vue'
 
 const props = defineProps({
