+++ i18n/en.json
@@ -394,7 +394,6 @@
   "ai.apiKey.description": "{provider} API Key is not set or invalid. Please enter a valid API key to use AI features.",
   "ai.apiKeyNotSet": "{provider} API Key is not set. Please ask your administrator to set it up",
   "ai.enterOpenAIAPIKey": "Enter OpenAI API Key",
-  "attachment.preview": "Preview",
   "auth.backToLogin": "Back to login",
   "auth.checkEmailForReset": "Check your email for the password reset link.",
   "auth.confirmPassword": "Confirm password",
@@ -690,6 +689,7 @@
   "globals.terms.description": "Description | Descriptions",
   "globals.terms.disabled": "Disabled",
   "globals.terms.draft": "Draft",
+  "globals.terms.download": "Download",
   "globals.terms.email": "Email | Emails",
   "globals.terms.enabled": "Enabled",
   "globals.terms.error": "Error | Errors",
@@ -838,8 +838,6 @@
   "globals.terms.white": "White",
   "globals.terms.workspace": "Workspace",
   "globals.terms.you": "You",
-  "imageLightbox.close": "Close",
-  "imageLightbox.download": "Download",
   "imageLightbox.next": "Next image",
   "imageLightbox.previous": "Previous image",
   "imageLightbox.resetZoom": "Reset zoom",
